LayoutTests/fast/dom/shadow/user-modify-inheritance.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="contenteditable-host1" contenteditable></div>
<div id="contenteditable-host2" contenteditable></div>
<div id="contenteditable-host3" contenteditable></div>
<div id="non-contenteditable-host1"></div>
<div id="non-contenteditable-host2"></div>
<div id="non-contenteditable-host3"></div>
<pre id="console"></pre>
<script>
description("Tests to ensure that a '-webkit-user-modify' CSS porperty is not inherited across shadow boundaries.");
if (window.testRunner)
    testRunner.dumpAsText();

function computedStyle(element, style) {
    var cs = window.getComputedStyle(element);
    if (!cs)
        return '(NA)';
    if (cs.getPropertyCSSValue(style))
        return cs.getPropertyCSSValue(style).cssText;
    return undefined;
}

function prepareNodeInShadowRoot(host, contentEditableAttributeValueForNode) {
    var shadowRoot = new WebKitShadowRoot(host);
    var nodeInShadow = document.createElement('div');
    if (contentEditableAttributeValueForNode != null)
        nodeInShadow.setAttribute('contenteditable', contentEditableAttributeValueForNode);
    shadowRoot.appendChild(nodeInShadow);
    document.body.offsetLeft;
    return nodeInShadow;
}

var userModifyPropertyName = '-webkit-user-modify';

shouldBeEqualToString('computedStyle(prepareNodeInShadowRoot(document.getElementById("contenteditable-host1"), null), userModifyPropertyName)', 'read-only');
shouldBeEqualToString('computedStyle(prepareNodeInShadowRoot(document.getElementById("contenteditable-host2"), "false"), userModifyPropertyName)', 'read-only');
shouldBeEqualToString('computedStyle(prepareNodeInShadowRoot(document.getElementById("contenteditable-host3"), "true"), userModifyPropertyName)', 'read-write');
shouldBeEqualToString('computedStyle(prepareNodeInShadowRoot(document.getElementById("non-contenteditable-host1"), null), userModifyPropertyName)', 'read-only');
shouldBeEqualToString('computedStyle(prepareNodeInShadowRoot(document.getElementById("non-contenteditable-host2"), "false"), userModifyPropertyName)', 'read-only');
shouldBeEqualToString('computedStyle(prepareNodeInShadowRoot(document.getElementById("non-contenteditable-host3"), "true"), userModifyPropertyName)', 'read-write');
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
